backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
httpx==0.25.0
python-dotenv==1.0.0
openai==1.2.4
pypdf==3.16.1 